split test_nirspec_ifu_spec3_regtest
